Add files via upload
I think these are all the files I've made changes too, again there is still a lot of stuff to add and fix and I'm working on it. The trouble I'm having now is finding a way to extract the piece object i added to the block class. If you notice whenever a piece is moved, The entire block is moved and an empty space is left that is because source is moving the entire block and replacing what ever it is moved to. The replacing is fine but we need it to replace the piece and not the block. If anybody can do some research on this and get back to me I'd appreciate it.