Additions
I tried not to change too much so that the code cant work for everybody, if you want to try moving the pieces with my code just uncomment the bottome half of the main and chessh controller files and comment out the top part and it should work i think. If it doesnt you may have to make the rest of the changes to the block file but just copy what i did with the first piece.